Rename builtin.function to builtins.functions
