+++ greenplumpython/builtins/functions.py
