+++ tests/test_dataframe.py
@@ -3,7 +3,7 @@ from os import environ
 import pytest
 
 import greenplumpython as gp
-from greenplumpython.builtin.function import generate_series
+from greenplumpython.builtins.functions import generate_series
 from tests import db
