+++ tests/test_func.py
@@ -4,7 +4,7 @@ from typing import Callable, List
 import pytest
 
 import greenplumpython as gp
-from greenplumpython.builtin.function import generate_series
+from greenplumpython.builtins.functions import generate_series
 from greenplumpython.func import AggregateFunction, NormalFunction
 from tests import db
